Create 2-2-1
Write a program that asks the user for an item type to buy, a number of items and a cost (remember to think about what data type should be used for each one). The program should then print a sentence repeating this information in a sentence as shown in the sample run.